requirements.txt
Flask>=0.10.1
boto>=2.31.1
pycrypto>=2.6.1
